removed reliance on all external gems (toml-rb, citrus)
