+++ kitchen.yml
@@ -15,37 +15,43 @@ verifier:
   name: inspec
 
 platforms:
-  # - name: amazonlinux
-  #   driver_config:
-  #     box: mvbcoding/awslinux
-  # - name: amazonlinux-2
-  # - name: centos-6
+  - name: amazonlinux
+    driver_config:
+      box: mvbcoding/awslinux
+  - name: amazonlinux-2
+  - name: centos-6
   - name: centos-7
     driver:
       box: bento/centos-7
-  # - name: centos-8
-  # - name: debian-9
-  # - name: debian-10
-  # - name: fedora-latest
-  # - name: opensuse-leap-15
-  # - name: ubuntu-16.04
-  # - name: ubuntu-18.04
+  - name: centos-8
+  - name: debian-9
+  - name: debian-10
+  - name: fedora-latest
+  - name: opensuse-leap-15
+  - name: ubuntu-16.04
+  - name: ubuntu-18.04
   - name: windows-2016
     driver:
       box: tas50/windows_2016
-  # - name: windows-2019
-  #   driver:
-  #     box: tas50/windows_2019
+  - name: windows-2019
+    driver:
+      box: tas50/windows_2019
 
 suites:
   # Testing Linux stuff here:
   - name: package
     named_run_list: package
+    verifier:
+      inspec_tests:
+        - test/integration/package
     excludes:
       - windows-2016
       - windows-2019
   - name: service
     named_run_list: service
+    verifier:
+      inspec_tests:
+        - test/integration/service
     excludes:
       - windows-2016
       - windows-2019
@@ -83,11 +89,17 @@ suites:
       - windows-2019
   - name: user-toml
     named_run_list: user_toml
+    verifier:
+      inspec_tests:
+        - test/integration/user_toml
     excludes:
       - windows-2016
       - windows-2019
   - name: config-chef-13
     named_run_list: config
+    verifier:
+      inspec_tests:
+        - test/integration/config
     excludes:
       - windows-2016
       - windows-2019
@@ -98,11 +110,17 @@ suites:
   # Testing windows stuff here:
   - name: win-package
     named_run_list: win_package
+    verifier:
+      inspec_tests:
+        - test/integration/win_package
     includes:
       - windows-2016
       - windows-2019
   - name: win-service
     named_run_list: win_service
+    verifier:
+      inspec_tests:
+        - test/integration/win_service
     includes:
       - windows-2016
       - windows-2019
@@ -116,11 +134,17 @@ suites:
       - windows-2019
   - name: win-config
     named_run_list: win_config
+    verifier:
+      inspec_tests:
+        - test/integration/win_config
     includes:
       - windows-2016
       - windows-2019
   - name: win-user-toml
     named_run_list: win_user_toml
+    verifier:
+      inspec_tests:
+        - test/integration/win_user_toml
     includes:
       - windows-2016
       - windows-2019
