+++ test/integration/config/default_spec.rb
@@ -14,8 +14,13 @@ describe command('hab -V') do
   its('exit_status') { should eq 0 }
 end
 
-splunkserviceapi = 'curl -v -H "Authorization: Bearer secret" http://localhost:9631/services/nginx/default | jq .cfg'
-describe json(command: splunkserviceapi) do
+describe file('/hab/svc/nginx/config/nginx.conf') do
+  its('content') { should include('worker_processes  2') }
+  its('content') { should include('keepalive_timeout  120') }
+end
+
+nginxserviceapi = 'curl -v -H "Authorization: Bearer secret" http://localhost:9631/services/nginx/default | jq .cfg'
+describe json(command: nginxserviceapi) do
   its(%w(http keepalive_timeout)) { should eq 120 }
   its(%w(http listen port)) { should eq 80 }
   its(['worker_processes']) { should eq 2 }
